added error handling in signup with thunk
