--- src/pages/Signup.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-function Signup() {
-  return <div></div>;
-}
-
-export default Signup;
